+++ src/components/ReplyDetails.tsx
@@ -5,6 +5,7 @@ import { Textbook } from "@/types/Textbook";
 import { MarkdownRenderer } from "./MarkdownRenderer";
 import { authClient } from "@/lib/auth-client";
 import { deleteReply } from "@/actions/reply";
+import { formatDate } from "@/lib/utils";
 import { 
   DropdownMenu, 
   DropdownMenuContent, 
@@ -28,27 +29,6 @@ export default function ReplyDetails(
   const { data: session } = authClient.useSession();
   
   const isAuthor = session?.user?.id === reply.author.id;
-  
-  const formatDate = (date: Date) => {
-    const now = new Date();
-    const diffInMs = now.getTime() - date.getTime();
-    const diffInMinutes = Math.floor(diffInMs / (1000 * 60));
-    const diffInHours = Math.floor(diffInMs / (1000 * 60 * 60));
-    const diffInDays = Math.floor(diffInMs / (1000 * 60 * 60 * 24));
-    
-    if (diffInMinutes < 1) return "Just now";
-    if (diffInMinutes < 60) return `${diffInMinutes}m ago`;
-    if (diffInHours < 24) return `${diffInHours}h ago`;
-    if (diffInDays < 7) return `${diffInDays}d ago`;
-    
-    return date.toLocaleDateString('en-US', {
-      year: 'numeric',
-      month: 'short',
-      day: 'numeric',
-      hour: '2-digit',
-      minute: '2-digit'
-    });
-  };
 
   const handleDelete = async () => {
     if (!confirm("Are you sure you want to delete this reply?")) return;
